quic: move the qlog tracer to a separate package
