+++ p2p/transport/quicutils/tracer_test.go
@@ -1,4 +1,4 @@
-package libp2pquic
+package quicutils
 
 import (
 	"bytes"
